refactor: move Checkout as pattern
