+++ package.json
@@ -89,6 +89,7 @@
     "qss": "~2.0.3",
     "react": "16.9.0",
     "react-codecopy": "~4.0.1",
+    "react-confetti": "~4.0.1",
     "react-dom": "16.9.0",
     "react-extras": "~0.9.0",
     "react-feather": "~2.0.3",
